requirements.txt
streamlit
requests
python-dotenv
streamlit-option-menu